# docs/swagger.yaml
swagger: "2.0"
info:
  title: Estate Agency API
  description: 'Введите токен в формате: "Bearer <token>" (включая слово Bearer и пробел)'
  version: 0.0.1
host: localhost:8080
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      summary: List users
      tags: [users]
      parameters:
        - name: limit
          in: query
          type: integer
          example: 10
        - name: offset
          in: query
          type: integer
          example: 0
        - name: ids
          in: query
          type: string
          example: "1,2,3"
        - name: email
          in: query
          type: string
          example: "user@example.com"
        - name: role
          in: query
          type: string
          example: "client"
        - name: is_active
          in: query
          type: boolean
          example: true
        - name: search
          in: query
          type: string
          example: "john"
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.ListUsersResponseDoc'
        "401":
          description: Требуется авторизация
        "500":
          description: Внутренняя ошибка сервера
  /users/register:
    post:
      summary: Register user
      tags: [users]
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userhandler.RegisterRequestDoc'
      responses:
        "201":
          description: Создано
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "409":
          description: Конфликт
        "500":
          description: Внутренняя ошибка сервера
  /users/login:
    post:
      summary: User login
      tags: [users]
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userhandler.LoginRequestDoc'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.LoginResponseDoc'
          examples:
            application/json:
              user:
                id: 1
                email: "user@example.com"
                first_name: "John"
                last_name: "Doe"
                role: "user"
                is_active: true
                created_at: "2025-11-29T12:00:00Z"
                updated_at: "2025-11-29T12:00:00Z"
              access_token: "eyJhbGciOiJIUzI1..."
              refresh_token: "dGhpc2lzYXJlZnJlc2h0b2tlbg=="
              expires_at: "2025-11-29T14:00:00Z"
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ErrorResponse'
  /users/{id}:
    get:
      summary: Get user
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "404":
          description: Не найдено
        "500":
          description: Внутренняя ошибка сервера
    delete:
      summary: Delete user
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "500":
          description: Внутренняя ошибка сервера
  /users/{id}/profile:
    patch:
      summary: Update user profile
      description: "Partially update user profile. Only provided fields will be changed. Requires owner or admin privileges. Returns the updated user." 
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userhandler.UpdateProfileRequestDoc'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "500":
          description: Внутренняя ошибка сервера
  /users/{id}/password:
    patch:
      summary: Change user password
      tags: [users]
      description: "Change user password. If performed successfully the updated user is returned. Admins can change other users' passwords."
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userhandler.ChangePasswordRequestDoc'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "500":
          description: Внутренняя ошибка сервера
  /users/{id}/role:
    patch:
      summary: Change user role
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userhandler.RoleRequestDoc'
      responses:
        "200":
          description: OK — returns the updated user with new role
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "500":
          description: Внутренняя ошибка сервера
  /users/{id}/active:
    patch:
      summary: Set user active status (toggle)
      tags: [users]
      description: "Toggle the user's active status and return the updated user. If the user was active, they become inactive and vice-versa."
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.PublicUserDoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "500":
          description: Внутренняя ошибка сервера
  /users/{id}/favorites:
    get:
      summary: List user favorites
      tags: [users]
      description: >-
        Возвращает список избранных объявлений указанного пользователя.
        Каждый элемент массива — объект PropertyDTO. Поля `images` и `image`
        присутствуют у каждого свойства: `image` содержит первую картинку, а
        `images` — список файлов (filename + data). Требуется авторизация.
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.ListFavoritesResponseDoc'
          examples:
            application/json:
              favorites:
                - id: 1
                  title: "Luxury Apartment"
                  property_description: "A spacious 3-bedroom apartment."
                  type_id: 2
                  transaction_type: "sale"
                  price: 350000
                  area: 120
                  property_address: "123 Main Street"
                  latitude: 40.7128
                  longitude: -74.006
                  city: "New York"
                  property_status: "available"
                  images:
                    - filename: "photo1.jpg"
                      data: "base64encodedstring=="
                    - filename: "photo2.jpg"
                      data: "base64encodedstring2=="
                  image:
                    filename: "photo1.jpg"
                    data: "base64encodedstring=="
                  created_at: "2025-11-29T12:00:00Z"
                  updated_at: "2025-11-29T12:00:00Z"
              total: 1
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "500":
          description: Внутренняя ошибка сервера
        "501":
          description: Не реализовано

  # -------------------- Tokens --------------------
...
definitions:
  propertyhandler.PropertyDTODoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      latitude: { type: number, example: 40.7128 }
      longitude: { type: number, example: -74.006 }
      city: { type: string, example: "New York" }
      property_status: { type: string, example: "available" }
      is_favorited: { type: boolean, example: false }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      updated_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      images:
        type: array
        items:
          $ref: '#/definitions/imagedto.ImageFileDoc'
      image:
        $ref: '#/definitions/imagedto.ImageFileDoc'
...
  imagedto.ImageFileDoc:
    type: object
    properties:
      filename: { type: string, example: "photo1.jpg" }
      data: { type: string, example: "base64encodedstring==", description: "Base64 encoded binary data of the image" }
...
  userhandler.RegisterRequestDoc:
    type: object
    required:
      - email
      - password
      - first_name
      - last_name
    properties:
      email: { type: string, example: "user@example.com" }
      password: { type: string, example: "secretpassword" }
      first_name: { type: string, example: "John" }
      last_name: { type: string, example: "Doe" }
      phone_number: { type: string, example: "+1234567890", nullable: true }
  userhandler.PublicUserDoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      email: { type: string, example: "user@example.com" }
      first_name: { type: string, example: "John" }
      last_name: { type: string, example: "Doe" }
      phone_number: { type: string, example: "+1234567890", nullable: true }
      role: { type: string, example: "user" }
      is_active: { type: boolean, example: true }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      updated_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
  userhandler.UpdateProfileRequestDoc:
    type: object
    properties:
      email: { type: string, example: "user@example.com" }
      first_name: { type: string, example: "John" }
      last_name: { type: string, example: "Doe" }
      phone_number: { type: string, example: "+1234567890", nullable: true }
  userhandler.ChangePasswordRequestDoc:
    type: object
    properties:
      current_password: { type: string, example: "oldpass123" }
      new_password: { type: string, example: "newpass456" }
  userhandler.RoleRequestDoc:
    type: object
    properties:
      role: { type: string, example: "admin" }
  userhandler.LoginRequestDoc:
    type: object
    properties:
      email: { type: string, example: "user@example.com" }
      password: { type: string, example: "secretpassword" }
  userhandler.LoginResponseDoc:
    type: object
    properties:
      user:
        $ref: '#/definitions/userhandler.PublicUserDoc'
      access_token: { type: string, example: "eyJhbGciOiJIUzI1..." }
      refresh_token: { type: string, example: "dGhpc2lzYXJlZnJlc2h0b2tlbg==" }
      expires_at: { type: string, format: date-time, example: "2025-11-29T14:00:00Z" }
  userhandler.ListFavoritesResponseDoc:
    type: object
    properties:
      favorites:
        type: array
        items:
          $ref: '#/definitions/propertyhandler.PropertyDTODoc'
      total: { type: integer, example: 2 }
  userhandler.ListUsersResponseDoc:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/userhandler.PublicUserDoc'
      total: { type: integer, example: 5 }
...
  ErrorResponse:
    type: object
    properties:
      code: { type: integer, example: 400 }
      message: { type: string, example: "Некорректные параметры запроса" }
      details:
        type: object
        additionalProperties:
          type: string
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header